+++ src/core/types.d.ts
@@ -1,0 +1,11 @@
+export type PjaxResources = {
+    eager: Array<string>;
+    lazy: Array<string>;
+};
+
+export type WorkerResponse = {
+    type: "eager" | "lazy" | "parse";
+    files: Array<string>;
+    requestUid: string | null;
+    pjaxFiles: PjaxResources;
+};